move ol_type to an external module, which can be easily customised: https://github.com/metanorma/isodoc/issues/653
